requirements.txt
Flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2 